+++ frontend/webapp/app/overview/sources/manage/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+import React from "react";
+
+export default function ManageSourcePage() {
+  return (
+    <>
+      <div>ManageSourcePage</div>
+    </>
+  );
+}
