+++ frontend/webapp/app/overview/sources/page.tsx
@@ -3,9 +3,5 @@ import { SourcesContainer } from "@/containers/overview";
 import React from "react";
 
 export default function SourcesOverviewPage() {
-  return (
-    <>
-      <SourcesContainer />
-    </>
-  );
+  return <SourcesContainer />;
 }
