+++ frontend/webapp/containers/overview/sources/sources.tsx
@@ -1,68 +1,24 @@
 "use client";
-import React, { useEffect, useState } from "react";
-import { NOTIFICATION, OVERVIEW, QUERIES } from "@/utils/constants";
+import React from "react";
+import { OVERVIEW, QUERIES, ROUTES } from "@/utils/constants";
 import { OverviewHeader } from "@/components/overview";
 import { SourcesContainerWrapper } from "./sources.styled";
-import { NewSourceFlow } from "./new.source.flow";
 import { ManageSources } from "./manage.sources";
-import { useNotification } from "@/hooks";
 import { useQuery } from "react-query";
 import { getSources } from "@/services";
+import { useRouter } from "next/navigation";
 
 export function SourcesContainer() {
-  const [displayNewSourceFlow, setDisplayNewSourceFlow] = useState<
-    boolean | null
-  >(null);
-  const { show, Notification } = useNotification();
-
-  const { data: sources, refetch } = useQuery(
-    [QUERIES.API_SOURCES],
-    getSources
-  );
-
-  useEffect(() => {
-    refetchSources();
-  }, [displayNewSourceFlow]);
-
-  async function refetchSources() {
-    if (displayNewSourceFlow === false) {
-      setTimeout(async () => {
-        refetch();
-      }, 1000);
-    }
-  }
-
-  function onNewSourceSuccess() {
-    setDisplayNewSourceFlow(false);
-    show({
-      type: NOTIFICATION.SUCCESS,
-      message: OVERVIEW.SOURCE_CREATED_SUCCESS,
-    });
-  }
-
-  function renderNewSourceFlow() {
-    return <NewSourceFlow onSuccess={onNewSourceSuccess} sources={sources} />;
-  }
-
-  function renderSources() {
-    return (
-      <ManageSources
-        setDisplayNewSourceFlow={setDisplayNewSourceFlow}
-        sources={sources}
-      />
-    );
-  }
+  const router = useRouter();
+  const { data: sources } = useQuery([QUERIES.API_SOURCES], getSources);
 
   return (
     <SourcesContainerWrapper>
-      <OverviewHeader
-        title={OVERVIEW.MENU.SOURCES}
-        onBackClick={
-          displayNewSourceFlow ? () => setDisplayNewSourceFlow(false) : null
-        }
+      <OverviewHeader title={OVERVIEW.MENU.SOURCES} />
+      <ManageSources
+        onAddClick={() => router.push(ROUTES.CREATE_SOURCE)}
+        sources={sources}
       />
-      {displayNewSourceFlow ? renderNewSourceFlow() : renderSources()}
-      <Notification />
     </SourcesContainerWrapper>
   );
 }
